LayoutTests/http/tests/navigation/page-cache-video-expected.txt

Tests that a page with a video element will pause after entering the page cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
video paused before page cache: false
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
PASS video paused in page cache
video paused after page cache: false
PASS successfullyParsed is true

TEST COMPLETE
